+++ src/opsem-interpreter/main.cc
@@ -3,6 +3,8 @@
 #include "parser.h"
 #include "visitors.h"
 
+#include "interpreter.h"
+
 #include <iostream>
 #include <string>
 #include <vector>
@@ -38,15 +40,7 @@ int main(int argc, const char** argv)
     l->accept(&printer);
   }
 
-  //Let's look at the functions first:
-  //cout << "THE FUNCTIONS" << endl;
-  //for (auto f: parser.functions) {
-  //  cout << f->function->name << endl;
-  //}
-  //cout << "THE CLASSES" << endl;
-  //for (auto c: parser.classes) {
-  //  cout << c->id->name << endl;
-  //}
+  Interpreter interp(parser);
 
   return 0;
 }
